[clean] Renamed files and moved function definitions for room update and delete from components to lib/client-actions
