+++ frontend/app/room/page.tsx
@@ -1,6 +1,4 @@
-import RoomCard from "@/components/RoomCard";
-
-export type Room = { id: number; name?: string };
+import RoomCard, { Room } from "@/components/room-card";
 
 async function getRooms(): Promise<Room[]> {
   const res = await fetch(`${process.env.API_URL}/room`, {
